src/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==21.4b1
click==7.1.2
dj-database-url==0.5.0
Django==3.1
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-utils-six==2.0
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
whitenoise==5.2.0
